modified:   lib/LEDStrip/LEDInfo.cpp 	modified:   lib/LEDStrip/LEDInfo.h It compiles! There's still much more to do.

+++ lib/LEDStrip/LEDInfo.h
@@ -8,7 +8,7 @@
 class LEDInfo {
 	private:
 		double* step;
-		double* deltas;
+		int* deltas;
 		double* carryOver;
 		void setStepValues(void);
